requirements.txt
Flask==2.0.3
telebot==0.0.4
pyTelegramBotAPI==3.8.3
gunicorn==19.9.0
Werkzeug==2.2.2
python-dotenv==1.0.1
pandas==1.0.3
pandas-gbq==0.17.4
google-cloud-bigquery==1.28.0
numpy==1.23.1
protobuf==3.20.1
google-api-core==1.33.2
requests==2.31.0


